renamed the tests to match
