+++ tests/auditory_tagging/stimulus_test_helpers.py
